[Jinsoo] Update

+++ Source/Project_LD/Public/Framework/Game/PS_Game.h
@@ -33,7 +33,7 @@ public:
 
 public:
 	void UpdateCurrentStats();
-	void UpdateExpValue(float InExp);
+	void UpdateExpValue(int InExp);
 	void IncreaseExp(float InExp);
 	void InitializeLocalPlayerState();
 	void InitializeLocalPlayerData();
